requirements.txt
aiogram
pillow
qrcode